+++ src/visitor.hpp
@@ -29,6 +29,7 @@ class CastExpr;
 class ArrayInitExpr;
 class ArrayIndexExpr;
 class EnumAccessExpr;
+class FieldAccessExpr;
 class EmptyStmt;
 class CompoundStmt;
 class LocalVarStmt;
@@ -68,6 +69,7 @@ class Visitor {
     virtual auto visit_array_init_expr(std::shared_ptr<ArrayInitExpr> array_init_expr) -> void = 0;
     virtual auto visit_array_index_expr(std::shared_ptr<ArrayIndexExpr> array_index_expr) -> void = 0;
     virtual auto visit_enum_access_expr(std::shared_ptr<EnumAccessExpr> enum_access_expr) -> void = 0;
+    virtual auto visit_field_access_expr(std::shared_ptr<FieldAccessExpr> field_access_expr) -> void = 0;
 
     virtual auto visit_empty_stmt(std::shared_ptr<EmptyStmt> empty_stmt) -> void = 0;
     virtual auto visit_compound_stmt(std::shared_ptr<CompoundStmt> compound_stmt) -> void = 0;